現状のままではノードの参照途切れとノード接続解除時の対応するFutureの削除に対応できないので構造を変更 親ノードの弱参照を保存するように、 親子各ノードとFuture及びPromiseをMap化

+++ ThreadPool/ThreadPool.cpp
@@ -199,7 +199,7 @@ std::optional<std::shared_ptr<ThreadPool>> ThreadPool::StealTaskFromOtherThread(
 	const thread::id currentThreadIdx = this_thread::get_id();
 
 	optional<thread::id> idx;
-	int32_t max = 0;
+	size_t max = 0;
 	for (auto& inThread : ms_Threads) {
 		if (inThread.first == currentThreadIdx)
 			continue;
